Restructure propagate methods and two-dimensional spectroscopy
Two-dimensional spectroscopy calculations require the ability to resume
propagation after the action of the dipole operator.  Propagate methods
which end by tracing out bath degrees of freedom preclude this.  Most
dynamics modules now have a propagate_full() function that does not
trace out the bath DOFs.

This commit also breaks backwards compatibility because the propagate()
methods (which do propagate_full() followed by an environmental trace)
no longer return the RDM in both the site and eigenstate basis (now just
in the site basis); transformation to a different basis is a
post-processing technique that it not the job of the dynamics routine.